Add student application feature
Logged out users can navigate to /applications/new and fill out a
student application form. Valid applications are saved and associated
with the user (also saved), displaying a confirmation screen and sending a
confirmation email to the student and a notification email with the
application to the admin.

Users created from applications are not activated (they can't log into
the app). This commit also simplifies them to have a single
name (instead of first and last), and password and password confirmation
are no longer required, since they aren't (and shouldn't be) part of the
application.

Infrastructure
--------------
In addition to the mailers and model tweaks/additions above, this commit
configures DatabaseCleaner in tests and sets up FactoryGirl.lint to run
before any specs (to catch invalid factories).