Add Course Invites and Registration features
Course Registration Invites
---------------------------

Adds ability to send course registration invites with
specially-generated JWTs to allow students to register without
activating their accounts (picking passwords). Currently, these invites
must be sent via the console. A later story may make it possible to do
invites through the app's web interface.

Infrastructure
--------------
- Course model
- Sets host for action mailer urls (including one that will cause the
  production server to break so I don't forget to set it there)

Allow applicants to pay for courses during registration
-------------------------------------------------------

This is a Stripe integration. The implementation allows for creating
Contract Plans and associating them with Courses (i.e. you can choose
different ways students can pay for courses). For installment plans,
Stripe subscriptions are created that automatically bill the student.
Contracts are used to track a student's payments, and in a future story,
I will implement webhooks for acting when a subscription is paid (by
reducing the contract balance and canceling the subscription if the
balance is 0).

This is currently only unit-tested. Integration testing may be done in a
future story, but the unit tests are pretty good.

Stripe recommends handling all error types it can send, but this initial
attempt is adequate for the time being -I'm ok with super edge cases
like the app crashing when a student tries to register and the Stripe
API is down.

To avoid any credit card data going to the app server at all, the credit
card data is instead sent directly to Stripe in an AJAX call that
returns a token if successful. This token can be safely sent to the app
server. As such, not all validation errors are necessarily displayed at
once.

Infrastructure Updates
----------------------
- lodash for use in javascript
- the rspec guard task is changed to use spring instead of bundler,
which makes loading the environment to run specs MUCH faster
- Adds InfoController to handle displaying custom messages when there
  isn't really a page the user should be on
- Adds display of flash in layout
- Contract Plans and Course Contract Plans to associate payment plans
with courses and stripe subscriptions (if need be)