Add domain-specific Guardfile
With the other Guardfile, rails restarts entirely whenever a file in lib/domain
changes. This is necessary when working with the app through the browser, but
makes auto-testing slow and painful when working with lib/domain files. This
domain-specific Guardfile addresses that by only watching the files in
lib/domain and spec/domain for changes and then running the relevant specs.

I initially tried to set it up as a separate group in the main Guardfile, but
when running all specs after `bundle exec guard -g domain`, every spec in the
spec dir (including rails ones) runs.